requirements.txt
torch
torchaudio
torchvision
transformers
accelerate
sentencepiece
streamlit
pillow
numpy
pandas
